+++ istio/datadog_checks/istio/constants.py
@@ -8,7 +8,7 @@ MESH_NAMESPACE = 'istio.mesh'
 PILOT_NAMESPACE = 'istio.pilot'
 GALLEY_NAMESPACE = 'istio.galley'
 CITADEL_NAMESPACE = 'istio.citadel'
-ISTIOD_NAMESPACE = 'istio.istiod' 
+ISTIOD_NAMESPACE = 'istio.istiod'
 # Would this be helpful to customers// break down istio and <components>?
 
 DEFAULT_METRIC_LIMIT = 0
