+++ istio/tests/test_isto_1_5.py
@@ -1,0 +1,3 @@
+# (C) Datadog, Inc. 2020-present
+# All rights reserved
+# Licensed under Simplified BSD License (see LICENSE)
